fix(agents): decode condition_score as int4 so listing details resolve

+++ src/agents/tools.rs
@@ -803,7 +803,7 @@ struct FullListingRow {
     title: String,
     category: String,
     brand: String,
-    condition_score: i64,
+    condition_score: i32,
     suggested_price_cny: i64,
     defects: String,
     description: Option<String>,
